+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 3.9.0 (in progress)
 ================
-* [#630](https://github.com/oshi/oshi/pull/630), [#640](https://github.com/oshi/oshi/pull/640): Add Sound Card list. - [@bilalAM](https://github.com/bilalAM).
+* [#630](https://github.com/oshi/oshi/pull/630), [#640](https://github.com/oshi/oshi/pull/640), [#645](https://github.com/oshi/oshi/pull/645): Add Sound Card list. - [@bilalAM](https://github.com/bilalAM).
 * [#636](https://github.com/oshi/oshi/pull/636): Catch exception when english counters missing. - [@dbwiddis](https://github.com/dbwiddis).
 * [#639](https://github.com/oshi/oshi/pull/639): Implement QueueLength metric in HWDiskStore. - [@Space2Man](https://github.com/Space2Man).
 * Your contribution here.